% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{calculate_activity}
\alias{calculate_activity}
\title{Calculate activity coefficients}
\usage{
calculate_activity(z, is, temp)
}
\arguments{
\item{z}{Charge of ions in the solution}

\item{is}{Ionic strength of the solution}

\item{temp}{Temperature of the solution in Celsius}
}
\value{
A numeric value for the activity coefficient.
}
\description{
This function calculates activity coefficients at a given temperature based on equation 5-43 from Davies (1967), Crittenden et al. (2012)
}
\examples{
calculate_activity(2, 0.1, 25)

}
