% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrosion_indices.R
\name{calculate_corrosion}
\alias{calculate_corrosion}
\title{Calculate six corrosion and scaling indices (AI, RI, LSI, LI, CSMR, CCPP)}
\usage{
calculate_corrosion(
  water,
  index = c("aggressive", "ryznar", "langelier", "ccpp", "larsonskold", "csmr")
)
}
\arguments{
\item{water}{Source water of class "water" created by \code{\link{define_water}}}

\item{index}{The indices to be calculated. 
Default calculates all six indices: "aggressive", "ryznar", "langelier", "ccpp", "larsonskold", "csmr"
CCPP may not be able to be calculated sometimes, so it may be advantageous to leave this out of the function to avoid errors}
}
\description{
\code{calculate_corrosion} takes an object of class "water" created by \code{\link{define_water}} and calculates 
corrosion and scaling indices.
}
\details{
Aggressive Index (AI) - the corrosive tendency of water and its effect on asbestos cement pipe
Ryznar Index(RI) - a measure of scaling potential
Langelier Saturation Index (LSI) - describes the potential for calcium carbonate scale formation
Larson-skold Index (LI) - describes the corrosivity towards mild steel
Chloride-to-sulfate mass ratio (CSMR) - indicator of galvanic corrosion for lead solder pipe joints
Calcium carbonate precipitation potential (CCPP) - a prediction of the mass of calcium carbonate that will precipitate at equilibrium
A positive CCPP value indicates the amount of CaCO3 (mg/L as CaCO3) that can be dissolved in the water. A negative CCPP
indicates how much CaCO3 will precipitate from solution.
}
\seealso{
\code{\link{define_water}}

 @examples
water <- define_water(ph = 8, temp = 25, alk = 200, ca_hard = 200, tds = 450, cl = 150, so4 = 200) %>%
 calculate_corrosion()
 
 water <- define_water(ph = 8, temp = 25, alk = 100, ca_hard = 50) %>%
calculate_corrosion(index = c("aggressive", "ccpp"))
}
