% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadsol_epa.R
\name{calculate_dic}
\alias{calculate_dic}
\title{Calc DIC (TOTCO3)}
\usage{
calculate_dic(water)
}
\arguments{
\item{water}{a data frame containing columns with all the parameters listed in \code{\link{define_water}}}
}
\description{
This function calcs dic
}
\examples{

example_df <- water_df \%>\% define_water_once()

}
\seealso{
\code{\link{define_water}}
}
