% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{calculate_dic_chain}
\alias{calculate_dic_chain}
\title{Calculate dissolved inorganic carbon (DIC) from total carbonate for a dataframe}
\usage{
calculate_dic_chain(df, input_water = "defined_water")
}
\arguments{
\item{df}{a data frame containing a water class column, which has already been computed using \link{define_water_chain}}

\item{input_water}{name of the column of Water class data to be used as the input for this function. Default is "defined_water".}
}
\value{
A numeric value for the calculated DIC.
}
\description{
This function takes a water class object defined by \code{\link{define_water}}
and outputs a DIC (mg/L). This function should be used when calculating DIC for multiple waters.
}
\examples{

example_dic <- water_df \%>\%
define_water_chain() \%>\%
calculate_dic_chain()


}
\seealso{
\code{\link{define_water}}
}
