% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_general.R
\name{calculate_hardness}
\alias{calculate_hardness}
\title{Hardness calculation}
\usage{
calculate_hardness(ca, mg, type = "total")

calculate_hardness(ca, mg, type = "total")
}
\arguments{
\item{ca}{Calcium concentration in mg/L as Ca}

\item{mg}{Magnesium concentration in mg/L as Mg}

\item{type}{"total" returns total hardness, "ca" returns calcium hardness}
}
\description{
This function takes Ca and Mg in mg/L and returns hardness in mg/L as CaCO3

This function takes Ca and Mg in mg/L and returns hardness in mg/L as CaCO3
}
\examples{
# Put example code here

# Put example code here

}
