% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdose_chloramine.R
\name{chemdose_chloramine}
\alias{chemdose_chloramine}
\title{Calculate Chlorine and Chloramine Concentrations with the Breakpoint Chlorination approach}
\source{
See references list at: \url{https://github.com/BrownandCaldwell-Public/tidywater/wiki/References}
}
\usage{
chemdose_chloramine(
  water,
  time,
  cl2,
  nh3,
  multi_cl_source = 1,
  cl_use_slot = FALSE,
  multi_nh3_source = 1,
  nh3_use_slot = FALSE
)
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}}

\item{time}{Reaction time (minutes).}

\item{cl2}{Applied chlorine dose (mg/L as Cl2).}

\item{nh3}{Applied ammonia dose (mg/L as N).}

\item{multi_cl_source}{# of chlorine source in use, default to 1 (1 or 2).}

\item{cl_use_slot}{If TRUE, uses free_chlorine slot in water instead of chlorine dose. Input for cl2 will be ignored. 
Default to FALSE, no effect if multi_cl_source is set to 2.}

\item{multi_nh3_source}{# of ammonia source in use, default to 1 (1 or 2).}

\item{nh3_use_slot}{If TRUE, uses tot_nh3 slot in water instead of ammonia dose. Input for nh3 will be ignored. 
Default to FALSE, no effect if multi_nh3_source is set to 2.}
}
\value{
A water class object with predicted Chlorine and Chloramine concentrations.
}
\description{
\code{simulate_breakpoint}, adopted from the U.S. EPA's Chlorine Breakpoint Curve Simulator, 
calculates chlorine and chlorinamine concentrations based on the two papers Jafvert & Valentine 
(Environ. Sci. Technol., 1992, 26 (3), pp 577-586) and Vikesland et al. (Water Res., 2001, 35 (7), pp 1766-1776).
Required arguments include an object of class "water" created by \code{\link{define_water}}, chlorine dose, and reaction time.
The function also requires additional water quality parameters defined in \code{\link{define_water}}
including temperature, pH, and alkalinity.
}
\details{
The function will calculate the Chlorine and Chloramine concentrations and update the "water"
class object proceed to the next steps of the treatment chain.
}
\examples{
example_breakpoint1 <- suppressWarnings(define_water(7.5, 20, 65, free_chlorine = 5,tot_nh3 = 1)) \%>\% 
  chemdose_chloramine(time = 40, cl2 = 2, nh3 = 1, cl_use_slot = TRUE, multi_nh3_source = 2)
example_breakpoint2 <- suppressWarnings(define_water(8, 20, 65)) \%>\% 
  chemdose_chloramine(time = 20, cl2 = 2, nh3 = 2, multi_cl_source = 2)

}
