% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{conv_chloramine}
\alias{conv_chloramine}
\title{Data frame of conversion factors for estimating DBP formation from chloramines}
\format{
A dataframe with 17 rows and 3 columns
\describe{
\item{ID}{abbreviation of dbp species}
\item{alias}{full name of dbp species}
\item{percent}{specifies the percent of DBP formation predicted from chloramines compared to chlorine, assuming the same chlorine dose applied}
}
}
\source{
U.S. EPA (2001), Table 5-10

See references list at: \url{https://github.com/BrownandCaldwell/tidywater/wiki/References}
}
\usage{
conv_chloramine
}
\description{
A dataset containing conversion factors for calculating DBP formation
}
\keyword{datasets}
