% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{convert_units}
\alias{convert_units}
\title{Unit Conversions}
\usage{
convert_units(value, formula, startunit = "mg/L", endunit = "M")
}
\arguments{
\item{value}{Value to be converted}

\item{formula}{Chemical formula of compound. Accepts compounds in \code{\link{mweights}} for conversions between g and mol or eq}

\item{startunit}{Units of current value, currently accepts g/L; g/L CaCO3; M; eq/L; and the same units with "m", "u", "n" prefixes}

\item{endunit}{Desired units, currently accepts same as start units}
}
\description{
This function takes a value and converts units based on compound name.
}
\examples{
# Put example code here

}
