% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_Water}
\alias{convert_Water}
\title{Convert Water}
\usage{
convert_Water(water)
}
\arguments{
\item{water}{A Water class vector}
}
\description{
Function to convert from Water class to a dataframe.
This is useful for one-off checks. It may be better to use fn_once
}
\examples{

example_df <- water_df \%>\%
define_water_chain \%>\%
mutate(to_dataframe  = map(defined_water, convert_Water)) \%>\%
unnest(to_dataframe) \%>\%
select(-defined_water)

}
\seealso{
\code{\link{define_water}}
}
