% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{correct_k}
\alias{correct_k}
\title{Correct acid dissociation constants}
\usage{
correct_k(water)
}
\arguments{
\item{water}{Defined water with values for temperature and ion concentrations}
}
\value{
A dataframe with equilibrium constants for co3, po4, so4, ocl, and nh4.
}
\description{
This function calculates the corrected equilibrium constant for temperature and ionic strength
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)
correct_k(water_defined)

}
