% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{define_water}
\alias{define_water}
\title{Create a water class}
\usage{
define_water(
  ph,
  temp,
  alk,
  tds,
  cond,
  tot_hard,
  ca_hard,
  na,
  k,
  cl,
  so4,
  tot_ocl = 0,
  po4 = 0
)
}
\arguments{
\item{ph}{water pH}

\item{temp}{Temperature in degree C}

\item{alk}{Alkalinity in mg/L as CaCO3}

\item{tds}{Total Dissolved Solids in mg/L}

\item{cond}{Electrical conductivity in}

\item{tot_hard}{Total hardness in mg/L as CaCO3}

\item{ca_hard}{Calcium hardness in mg/L as CaCO3}

\item{na}{Sodium in mg/L Na+}

\item{k}{Potassium in mg/L K+}

\item{cl}{Chloride in mg/L Cl-}

\item{so4}{Sulfate in mg/L SO42-}

\item{tot_ocl}{Chlorine in mg/L as ??}

\item{po4}{Phosphate in mg/L as PO4}
}
\description{
\code{define_water} takes user-defined water quality parameters and creates an object of class "water" that
forms the input and output of all other tidywater water treatment modelling functions. The function calculates
ionic strength and the carbonate balance, converts all ion concentrations to mol/L, and corrects for ionic
activity coefficients.
}
\details{
\code{define_water} calculates ionic strength and activity coefficients using the following established water chemistry
equations:
Ionic strength (if TDS provided): MWH equation 5-38
Ionic strength (if electrical conductivity provided): Snoeyink & Jenkins 1980
Ionic strength (from ion concentrations): Lewis and Randall (1921), MWH equation 5-37
Temperature correction of dielectric constant (relative permittivity): Harned and Owen (1958), MWH equation 5-45.
Activity coefficients: Davies equation (1967), MWH equation 5-43
Activity coefficient constant A: Stumm and Morgan (1996), Trussell (1998), MWH equation 5-44
}
\examples{
define_water(ph=7,temp=20,alk=100,tds=10)

}
