% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{discons}
\alias{discons}
\title{Dissociation constants for weak acids/bases}
\format{
A dataframe with one row and one column per constant
}
\usage{
discons
}
\description{
A dataset containing the dissociation constants for carbonate, sulfate, phosphate, and hypochlorite
}
\keyword{datasets}
