% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_cu.R
\name{dissolve_cu_once}
\alias{dissolve_cu_once}
\title{Calculate Dissolved Copper Concentration}
\usage{
dissolve_cu_once(df, input_water = "defined_water")
}
\arguments{
\item{df}{a data frame containing a water class column, which has already been computed using \link{define_water_chain}}

\item{input_water}{name of the column of Water class data to be used as the input for this function. Default is "defined_water".}
}
\value{
\code{dissolve_cu_once} returns a data frame containing the original data frame and a column for dissolved copper in mg/L.
}
\description{
Calculate Dissolved Copper Concentration
}
\examples{
library(dplyr)
cu_calc <- water_df \%>\%
  define_water_chain() \%>\%
  dissolve_cu_once()

}
