% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadsol_epa.R
\name{dissolve_pb}
\alias{dissolve_pb}
\title{Simulate contributions of various lead solids to total soluble lead}
\usage{
dissolve_pb(water)
}
\arguments{
\item{water}{a data frame containing columns with all the parameters listed in \code{\link{define_water}}}
}
\description{
This function takes a water data frame defined by \code{\link{define_water}} 
and outputs a dataframe. Lead solid solubility is calculated based on controlling solid (Pb_2_plus, M).
Total dissolved lead species (tot_dissolved_pbm, M) are calculated based on lead complex calculations.
}
\details{
The solid with lowest solubility will form the lead scale (controlling lead solid)

Make sure that total dissolved solids, conductivity, or 
ca, na, cl, so4 are used in define_water so that an ionic strength is calculated.

Code is from EPA's TELSS lead solubility dashboard https://github.com/USEPA/TELSS/blob/main/app.R
}
\examples{

example_pb <- define_water(ph = 7.5, temp = 25, alk = 93, cl = 240, tot_po4 = 0, so4 = 150, tds = 200) \%>\%
 dissolve_pb()

}
\seealso{
\code{\link{define_water}}
}
