% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid Base Equilibrium.R
\name{dose}
\alias{dose}
\title{Chemical Dose Function}
\usage{
dose(
  water,
  hcl = 0,
  h2so4 = 0,
  h3po4 = 0,
  naoh = 0,
  na2co3 = 0,
  nahco3 = 0,
  caoh2 = 0,
  mgoh2 = 0,
  cl2 = 0,
  naocl = 0,
  caocl2 = 0
)
}
\arguments{
\item{water}{Source water vector created by link function here}

\item{hcl}{Hydrochloric acid}

\item{h2so4}{Sulfuric acid = H2SO4 -> 2H + SO4}

\item{naoh}{Caustic = NaOH -> Na + OH}

\item{na2co3}{Soda Ash = Na2CO3 -> 2Na + CO3}

\item{nahco3}{description}

\item{caoh2}{Lime = Ca(OH)2 -> Ca + 2OH}

\item{mgoh2}{description}

\item{cl2}{Chlorine gas = Cl2(g) + H2O -> HOCl + H + Cl}

\item{naocl}{Sodium Hypochlorite = NaOCl -> Na + OCl}

\item{caocl2}{description}

\item{h3op4}{Phosphoric acid}
}
\description{
This function takes chemical doses and a source water vector and outpus a new sourcewater vector with updated ion balance an pH.
Units of all chemical additions in mg/L.
Returns dataframe of dosed water quality.
}
\examples{
# Put example code here

}
