% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_base_equilibirum.R
\name{dose_target}
\alias{dose_target}
\title{Target Chemical Dose Function}
\usage{
dose_target(water, target_ph, chemical)
}
\arguments{
\item{water}{Source water data frame created by \code{\link{define_water}}}

\item{target_ph}{The final pH to be achieved after the specified chemical is added.}

\item{chemical}{The chemical to be added. Current supported chemicals include: caustic soda (NaOH), lime (Ca(OH2)), magnesium hydroxide (Mg(OH)2), and CO2.}
}
\description{
This function calculates the required amount of a chemical to dose based on a target pH and existing water quality.
Returns numeric value for dose in mg/L. Uses optimize on the dose_chemical function.
}
\examples{
# Put example code here

}
\seealso{
\code{\link{define_water}}
}
