% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{edwardscoeff}
\alias{edwardscoeff}
\title{Data frame of Edwards model coefficients}
\format{
A dataframe with 5 rows and 7 columns:
\describe{
\item{ID}{Coefficient type}
\item{x3}{x3 parameter}
\item{x2}{x2 parameter}
\item{x1}{x1 parameter}
\item{k1}{k1 parameter}
\item{k2}{k2 parameter}
\item{b}{b parameter}
}
}
\source{
Edwards (1997) Table 2.

See references list at: \url{https://github.com/BrownandCaldwell-Public/tidywater/wiki/References}
}
\usage{
edwardscoeff
}
\description{
A dataset containing coefficients from the Edwards (1997) model for coagulation TOC removal.
}
\keyword{datasets}
