% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gacrun_toc.R
\name{gacrun_toc}
\alias{gacrun_toc}
\title{Calculate TOC Concentration in GAC system}
\source{
See references list at: \url{https://github.com/BrownandCaldwell-Public/tidywater/wiki/References}

Zachman and Summers (2010)

BC WTP Model v. 2.0 Manual (2001)
}
\usage{
gacrun_toc(water, ebct = 10, model, media_size = "12x40")
}
\arguments{
\item{water}{Source water object of class "water" created by \link{define_water}}

\item{ebct}{Empty bed contact time (minutes). Model results are valid for 10 or 20 minutes.}

\item{model}{Specifies which GAC TOC removal model to apply. Options are Zachman and WTP.}

\item{media_size}{Size of GAC filter mesh. Model includes 12x40 and 8x30 mesh sizes.}
}
\value{
\code{gacrun_toc} returns a data frame with bed volumes and breakthrough TOC values.
}
\description{
Calculates TOC concentration after passing through GAC treatment according to the model developed in
"Modeling TOC Breakthrough in Granular Activated Carbon Adsorbers" by Zachman and Summers (2010), or the BC WTP Model v. 2.0 Manual.

Water must contain DOC or TOC value.
}
\details{
GAC model for TOC removal

The function will calculate bed volumes and normalized TOC breakthrough (TOCeff/TOCinf) given model type.
\subsection{should I include model limitations in the details?}{
}
}
\examples{
water <- define_water(ph = 8, toc = 2.5, uv254 = .05, doc = 1.5) \%>\%
  gacrun_toc(media_size = "8x30", ebct = 20, model = "Zachman")

}
