% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{leadplotcoeffs}
\alias{leadplotcoeffs}
\title{Data frame of lead solubility at various pH and DIC values}
\format{
A dataframe with 434 rows and 3 columns
\describe{
\item{leadplotph}{A range of pH values from 7 to 10}
\item{leadplotdic}{A range of DIC values from 2 to 58 mg/L}
\item{leadplotsol}{Lead solubility values corresponding to the given pH and DIC}
}
}
\source{
NEED TO FILL IN

See references list at: \url{https://github.com/BrownandCaldwell-Public/tidywater/wiki/References}
}
\usage{
leadplotcoeffs
}
\description{
A dataset containing dissolved lead, pH, and DIC
}
\keyword{datasets}
