% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{leadplottransition}
\alias{leadplottransition}
\title{Data frame of the pH-DIC transition line for lead contour plot}
\format{
A dataframe with 19 rows and 2 columns
\describe{
\item{ph}{A range of pH values between 7.1 and 8.649}
\item{dic}{A range of DIC values from 0.8 to 58 mg/L}
}
}
\source{
NEED TO FILL IN

See references list at: \url{https://github.com/BrownandCaldwell-Public/tidywater/wiki/References}
}
\usage{
leadplottransition
}
\description{
A dataset containing pH and DIC for the contour plot transition line
}
\keyword{datasets}
