% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{mweights}
\alias{mweights}
\title{Molar weights of relevant compounds}
\format{
A dataframe with one row and one column per compound
}
\usage{
mweights
}
\description{
A dataset containing the molar weights of several compounds in g/mol
}
\keyword{datasets}
