% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disinfection_ct.R
\name{ozonate_ct}
\alias{ozonate_ct}
\title{Determine disinfection credit from ozone.}
\source{
USEPA (2020) Equation 4-4 through 4-7
https://www.epa.gov/system/files/documents/2022-02/disprof_bench_3rules_final_508.pdf

See references list at: \url{https://github.com/BrownandCaldwell/tidywater/wiki/References}
}
\usage{
ozonate_ct(water, time, dose, kd)
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}. Water must include ph and temp}

\item{time}{Retention time of disinfection segment in minutes.}

\item{dose}{Ozone dose in mg/L. This value can also be the y intercept of the decay curve (often slightly lower than ozone dose.)}

\item{kd}{First order decay constant.}
}
\description{
This function takes a water defined by \code{\link{define_water}} and the first order decay curve parameters
from an ozone dose and outputs a dataframe of acutal CT, and log removal for giardia, virus, and crypto
}
\details{
First order decay curve for ozone has the form: `residual = dose * exp(kd*time)`. kd should be a negative number.
Actual CT is an integration of the first order curve. The first 30 seconds are removed from the integral to account for
instantaneous demand.
}
\examples{

example_ct <- define_water(ph = 7.5, temp = 25) \%>\%
  ozonate_ct(time = 10, dose = 2, kd = -0.5)
}
\seealso{
\code{\link{define_water}}
}
