% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{pactoccoeffs}
\alias{pactoccoeffs}
\title{Data frame of PAC TOC model coefficients}
\format{
A dataframe with 4 rows and 3 columns
\describe{
\item{pactype}{Specifies PAC type}
\item{A}{Constant in the PAC model}
\item{a}{Coefficient in PAC model, associated with DOC0}
\item{b}{Coefficient in PAC model, associated with dose}
\item{c}{Coefficient in PAC model, associated with time}
}
}
\source{
Cho (2007)
}
\usage{
pactoccoeffs
}
\description{
A dataset containing coefficients for calculating PAC TOC removal
}
\keyword{datasets}
