% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_base_equilibirum.R
\name{plot_ions}
\alias{plot_ions}
\title{Ion Summary Plot}
\usage{
plot_ions(water, title = "")
}
\arguments{
\item{water}{Source water vector created by link function here}

\item{title}{Optional plot title}
}
\description{
This function takes a water data frame defined by \code{\link{define_water}} and outputs an ion balance plot.
}
\examples{
# Put example code here

}
