% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plot_lead}
\alias{plot_lead}
\title{Create dissolved lead and DIC contour plot given input data frame}
\usage{
plot_lead(df)
}
\arguments{
\item{df}{Source data as a data frame. Must have columns for historical pH and DIC data.}
}
\value{
A ggplot object displaying a contour plot of dissolved lead, pH, and DIC
}
\description{
This function takes a data frame and outputs a contour plot of dissolved lead and DIC plot.
}
\examples{
historical <- data.frame(ph = c(7.7, 7.86, 8.31, 7.58, 7.9, 8.06, 7.95, 8.02, 7.93, 7.61),
                dic = c(14.85813772, 16.40951309, 16.47773344, 16.6332028, 16.86320599, 16.93871268, 17.05489053, 17.22845295, 17.32943129, 17.34448707))
plot_lead(historical)

}
