% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid Base Equilibrium.R
\name{plot_wq}
\alias{plot_wq}
\title{Water Summary Plot}
\usage{
plot_wq(water, title = "")
}
\arguments{
\item{water}{Source water vector created by link function here}

\item{title}{Optional plot title}
}
\description{
This function takes a water data frame defined by 'define_water' and outputs an ion balance plot.
}
\examples{
# Put example code here

}
