% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pluck_water}
\alias{pluck_water}
\title{Pluck out a single parameter from a water class object}
\usage{
pluck_water(df, input_water = "defined_water", parameter, output_column = NULL)
}
\arguments{
\item{df}{a data frame containing a column, defined_water, which has already
been computed using \code{\link{define_water}}}

\item{input_water}{name of the column of Water class data to be used as the input for this function.}

\item{parameter}{water class attribute to view outside the water column}

\item{output_column}{name of output column storing the plucked variable's values}
}
\description{
This function plucks a selected parameter from a column of water class objects.
To view multiple parameters, please use one of the "fn_once" functions or \code{\link{convert_water}}.
}
\examples{

library(dplyr)
library(furrr)
library(purrr)
library(tidyr)

pluck_example <- water_df \%>\%
  define_water_chain() \%>\%
  pluck_water(parameter = "tot_co3")

pluck_example <- water_df \%>\%
  define_water_chain() \%>\%
  balance_ions_chain() \%>\%
  pluck_water(parameter = "na", output_column = "defined_na") \%>\%
  pluck_water(input_water = "balanced_water", parameter = "na", output_column = "balanced_na")

plan(multisession)
pluck_example <- water_df \%>\%
  define_water_chain() \%>\%
  pluck_water(parameter = "ph")

# Optional: explicitly close multisession processing
plan(sequential)
}
\seealso{
\code{\link{convert_water}}
}
