% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadsol_epa.R
\name{simulate_solubility}
\alias{simulate_solubility}
\title{Calc lead sol}
\usage{
simulate_solubility(water, IS)
}
\arguments{
\item{water}{a data frame containing columns with all the parameters listed in \code{\link{define_water}}}
}
\description{
This function calcs leadsol
}
\examples{

example_df <- water_df \%>\% define_water_once()

}
\seealso{
\code{\link{define_water}}
}
