% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{solve_ph_rust_inner}
\alias{solve_ph_rust_inner}
\title{Internal Rust function to solve pH for water chemistry equilibrium}
\usage{
solve_ph_rust_inner(
  temp,
  ionic_strength,
  kw,
  tot_po4,
  tot_co3,
  tot_ocl,
  tot_nh3,
  tot_ch3coo,
  h2po4_i,
  hpo4_i,
  po4_i,
  ocl_i,
  nh4_i,
  ch3coo_i,
  carbonate_alk_eq,
  oh_i,
  h_i,
  so4_dose,
  na_dose,
  ca_dose,
  mg_dose,
  cl_dose,
  mno4_dose,
  no3_dose
)
}
\arguments{
\item{temp}{Temperature in Celsius}

\item{ionic_strength}{Ionic strength in M (optional, use NULL if not available)}

\item{kw}{Water dissociation constant}

\item{tot_po4}{Total phosphate concentration}

\item{tot_co3}{Total carbonate concentration}

\item{tot_ocl}{Total hypochlorite concentration (free chlorine)}

\item{tot_nh3}{Total ammonia concentration}

\item{tot_ch3coo}{Total acetate concentration}

\item{h2po4_i}{Initial H2PO4- concentration}

\item{hpo4_i}{Initial HPO4^2- concentration}

\item{po4_i}{Initial PO4^3- concentration}

\item{ocl_i}{Initial OCl- concentration}

\item{nh4_i}{Initial NH4+ concentration}

\item{ch3coo_i}{Initial CH3COO- concentration}

\item{carbonate_alk_eq}{Carbonate alkalinity in equivalents}

\item{oh_i}{Initial OH- concentration}

\item{h_i}{Initial H+ concentration}

\item{so4_dose}{Sulfate dose}

\item{na_dose}{Sodium dose}

\item{ca_dose}{Calcium dose}

\item{mg_dose}{Magnesium dose}

\item{cl_dose}{Chloride dose}

\item{mno4_dose}{Permanganate dose}

\item{no3_dose}{Nitrate dose}
}
\description{
This function takes individual parameters and performs the pH calculation.
It should be called from the R wrapper function.
}
