% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvecost.R
\name{solvecost_labor}
\alias{solvecost_labor}
\title{Determine labor cost}
\usage{
solvecost_labor(fte, cost, time = "day")
}
\arguments{
\item{fte}{Number of FTEs. Can be decimal.}

\item{cost}{$/year per FTE}

\item{time}{Desired output units, one of c("day", "month", "year"). Defaults to "day".}
}
\value{
A numeric value for labor $/time.
}
\description{
This function takes number of FTE and annual $/FTE and determines labor cost
}
\examples{
laborcost <- solvecost_labor(1.5, 50000)

library(dplyr)
cost_data <- tibble(
  fte = seq(1, 10, 1)
) \%>\%
  mutate(costs = solvecost_labor(fte = fte, cost = .08))

}
