% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvecost.R
\name{solvecost_power}
\alias{solvecost_power}
\title{Determine power cost}
\usage{
solvecost_power(power, utilization = 100, cost, time = "day")
}
\arguments{
\item{power}{Power consumed in kW}

\item{utilization}{Amount of time equipment is running in percent. Defaults to continuous.}

\item{cost}{Power cost in $/kWhr}

\item{time}{Desired output units, one of c("day", "month", "year"). Defaults to "day".}
}
\value{
A numeric value for power, $/time.
}
\description{
This function takes kW, \% utilization, $/kWhr and determines power cost.
}
\examples{
powercost <- solvecost_power(50, 100, .08)

library(dplyr)
cost_data <- tibble(
  power = seq(10, 50, 10),
  utilization = 80
) \%>\%
  mutate(costs = solvecost_power(power = power, utilization = utilization, cost = .08))

}
