% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{solvedose_ph_once}
\alias{solvedose_ph_once}
\title{Dose Target Once}
\usage{
solvedose_ph_once(
  df,
  input_water = "defined_water",
  output_column = "dose_required",
  target_ph = NULL,
  chemical = NULL
)
}
\arguments{
\item{df}{a data frame containing a column, defined_water, which has already
been computed using \code{\link{define_water}}, and a column named for each of the chemicals being dosed}

\item{input_water}{name of the column of Water class data to be used as the input for this function. Default is "defined_water".}

\item{output_column}{name of the output column storing doses in mg/L. Default is "dose_required".}

\item{target_ph}{set a goal for pH using the function argument or a data frame column}

\item{chemical}{select the chemical to be used to reach the desired pH using function argument or data frame column}
}
\description{
This function allows \code{\link{solvedose_ph}} to be added to a piped data frame. 
Its output is a chemical dose in mg/L.
}
\details{
The data input comes from a Water class column, as initialized in \code{\link{define_water}} or \code{\link{balance_ions}}.

If the input data frame has a column(s) named "target_ph" or "chemical", the function will use the column(s) in addition to the 
as a function argument. If these columns aren't present, specify "target_ph" or "chemical" as function arguments.
The chemical names must match the chemical names as displayed in \code{\link{solvedose_ph}}. 
To see which chemicals can be dosed, see \code{\link{solvedose_ph}}. 

The input data frame requires a column, "chemical", indicating which chemical should be used to reach the target pH
The input data frame also requires a column, "target_ph", indicating the final pH after the selected chemical is added
The column names must match the chemical names as displayed in \code{\link{solvedose_ph}}. 
To see which chemicals can be passed into the function, see \code{\link{solvedose_ph}}.
}
\examples{

example_df <- water_df \%>\% 
define_water_chain() \%>\% 
balance_ions_chain() \%>\%
mutate(target_ph = 10,
      chemical = rep(c("naoh", "mgoh2"), 6)) \%>\%
solvedose_ph_once()

example_df <- water_df \%>\% 
define_water_chain() \%>\% 
balance_ions_chain() \%>\%
solvedose_ph_once(input_water = "balanced_Water", target_ph = 8.8, chemical = "naoh")


example_df <- water_df \%>\% 
define_water_chain() \%>\% 
mutate(target_ph = seq(9, 10.1, .1)) \%>\%
solvedose_ph_once(chemical = "naoh")


}
\seealso{
\code{\link{solvedose_ph}}
}
