% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid Base Equilibrium.R
\name{sourceWater}
\alias{sourceWater}
\title{Source water vector}
\usage{
sourceWater(ph, temp, alk, tot_hard, c_hard, na, k, cl, so4, tot_ocl = 0)
}
\arguments{
\item{temp}{Temperature in degree C}

\item{alk}{Alkalinity in mg/L as CaCO3}

\item{na}{Sodium in mg/L Na+}

\item{cl}{Chloride in mg/L Cl-}

\item{so4}{Sulfate in mg/L SO42-}

\item{tot_ocl}{Chlorine in mg/L as ??}

\item{hardT}{Total hardness in mg/L as CaCO3}

\item{hardC}{Calcium hardness in mg/L as CaCO3}

\item{ca}{Calcium in mg/L as CaCO3}

\item{mg}{Magnesium in mg/L Mg2+}
}
\description{
This function takes water quality parameters and creates a standard vector that is the input and output of all pH functions.
Carbonate balance is calculated and units are converted to mol/L
}
\examples{
# Put example code here

}
