% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrosion_indices.R
\name{summarise_corrosion}
\alias{summarise_corrosion}
\title{Create summary table of corrosion indices from a water class}
\usage{
summarise_corrosion(water)
}
\arguments{
\item{water}{Source water vector created by \code{\link{define_water}}.}
}
\description{
This function takes a water data frame defined by \code{\link{define_water}} and outputs a formatted summary table of
corrosion indices calculated by \code{\link{calculate_corrosion}}.
}
\examples{
water <- define_water(ph = 8, temp = 25, alk = 200, ca_hard = 200, tds = 576) \%>\%
  calculate_corrosion()
summarise_corrosion(water)

}
