% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summarise_dbp}
\alias{summarise_dbp}
\title{Create DBP summary table from water class}
\usage{
summarise_dbp(water)
}
\arguments{
\item{water}{Source water vector created by \code{\link{chemdose_dbp}}.}
}
\description{
This function takes a water data frame defined by \code{\link{chemdose_dbp}} and and outputs a formatted summary table of
modeled DBP concentrations.
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)
summarise_dbp(water_defined)

}
