% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summarise_wq}
\alias{summarise_wq}
\title{Create summary table from water class}
\usage{
summarise_wq(water)
}
\arguments{
\item{water}{Source water vector created by \code{\link{define_water}}.}
}
\description{
This function takes a water data frame defined by \code{\link{define_water}} and outputs a formatted summary table of
general water quality parameters and major ions.
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)
summarise_wq(water_defined)

}
