% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summarize_wq}
\alias{summarize_wq}
\title{Water Summary Table}
\usage{
summarize_wq(water)
}
\arguments{
\item{water}{Source water vector created by link function here}
}
\description{
This function takes a water data frame defined by \code{\link{define_water}} and outputs a formatted summary table.
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)
summarize_wq(water_defined)

}
