% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{toc_compliance_table}
\alias{toc_compliance_table}
\title{Data frame of TOC removal requirements per Stage 1 DBP Rule}
\format{
A dataframe with 9 rows and 5 columns
\describe{
\item{toc_min}{Minimum raw water TOC for compliance}
\item{toc_max}{Maximum raw water TOC for compliance}
\item{alk_min}{Minimum raw water alkalinity for compliance}
\item{alk_max}{Maximum raw water alkalinity for compliance}
\item{required_compliance}{Required percent TOC removal based on raw TOC andn alklalinity range}
}
}
\source{
Stage 1 Disinfection Byproduct Rule (1998)
}
\usage{
toc_compliance_table
}
\description{
A dataset containing required TOC removal based on raw alkalinity and TOC
}
\keyword{datasets}
