% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{water_df}
\alias{water_df}
\title{Data frame of water quality parameters}
\format{
A dataframe with 12 rows and 11 columns:
\describe{
\item{ph}{pH in standard units (SU)}
\item{temp}{Temperature in degree C}
\item{alk}{Alkalinity in mg/L as CaCO3}
\item{tot_hard}{Total hardness in mg/L as CaCO3}
\item{ca_hard}{Calcium hardness in mg/L as CaCO3}
\item{na}{Sodium in mg/L Na+}
\item{k}{Potassium in mg/L K+}
\item{cl}{Chloride in mg/L Cl-}
\item{so4}{Sulfate in mg/L SO42-}
\item{tot_ocl}{Total chlorine in mg/L as Cl2}
\item{tot_po4}{Total phosphate in mg/L as PO42-}
}
}
\source{
Fabricated for use in examples.
}
\usage{
water_df
}
\description{
A dataset containing fabricated water quality to use as tidywater inputs.
Parameters are set to reasonable water quality ranges. Parameters are as follows:
}
\keyword{datasets}
