% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{water_tomg}
\alias{water_tomg}
\title{Convert primary units in a water from M to mg/L}
\usage{
water_tomg(water)
}
\arguments{
\item{water}{Water created with define_water, which currently has primary_unit == "M"}
}
\description{
This function takes a water defined by \code{\link{define_water}} and converts the following slots from
M to mg/L: na, ca, mg, k, cl, so4, hco3, co3, h2po4, hpo4, po4, ocl, bro3, f, fe, al.  These slots are converted to
ug/L: br, mn.  All other values remain unchanged.
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1) \%>\%
  water_tomg()

}
