% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{water_tomol}
\alias{water_tomol}
\title{Convert primary units in a water from mg/L to M}
\usage{
water_tomol(water)
}
\arguments{
\item{water}{Water created with define_water and converted with water_tomg, which currently has primary_unit == "mg/L"}
}
\description{
This function takes a water defined by \code{\link{define_water}} and converted to mg/L by
\code{\link{water_tomg}} and converts the following slots back to M: na, ca, mg, k, cl, so4, hco3, co3, h2po4, hpo4,
po4, ocl, bro3, f, fe, al, br, mn.  All other values remain unchanged.
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1) \%>\%
  water_tomg() \%>\%
  water_tomol()

}
