% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid Base Equilibrium.R
\name{waterplot}
\alias{waterplot}
\title{Water Summary Plot}
\usage{
waterplot(water, title = "")
}
\arguments{
\item{water}{Source water vector created by link function here}

\item{title}{Optional plot title}
}
\description{
This function takes a source water vector and outputs an ion balance plot.
}
\examples{
# Put example code here

}
