% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid Base Equilibrium.R
\name{watersummary}
\alias{watersummary}
\title{Water Summary Table}
\usage{
watersummary(water)
}
\arguments{
\item{water}{Source water vector created by link function here}
}
\description{
This function takes a source water vector and outputs a formatted summary table.
}
\examples{
# Put example code here

}
