% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_general.R
\name{balance_ions}
\alias{balance_ions}
\title{Ion balance a water}
\usage{
balance_ions(water)
}
\arguments{
\item{water}{Water created with define_water, which may have some ions set to 0 when unknown}
}
\description{
This function takes a water defined by \code{\link{define_water}} and balances charge. If more cations are needed, sodium
will be added, unless a number for sodium is already provided and potassium is 0, then it will add potassium. Similarly,
anions are added using chloride, unless sulfate is 0.
}
\examples{
# Put example code here

}
