% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{balance_ions_chain}
\alias{balance_ions_chain}
\title{Apply balance_ions within a dataframe and output a column of water class to be chained to other tidywater functions}
\usage{
balance_ions_chain(
  df,
  input_water = "defined_water",
  output_water = "balanced_water"
)
}
\arguments{
\item{df}{a data frame containing a column, defined_water, which has already been computed using \code{\link{define_water}}}

\item{input_water}{name of the column of Water class data to be used as the input for this function. Default is "defined_water".}

\item{output_water}{name of the output column storing updated parameters with the class, Water. Default is "balanced_water".}
}
\description{
This function allows \code{\link{balance_ions}} to be added to a piped data frame.
Its output is a "water" class, and can therefore be used with "downstream" tidywater functions.
}
\examples{
library(purrr)
library(furrr)
library(tidyr)
library(dplyr)

example_df <- water_df \%>\%
define_water_chain() \%>\%
balance_ions_chain() \%>\%
chemdose_ph_chain(naoh = 5)

example_df <- water_df \%>\%
define_water_chain() \%>\%
balance_ions_chain(output_water = "balanced ions, balanced life") \%>\%
chemdose_ph_chain(input_water = "balanced ions, balanced life", naoh =5)

}
\seealso{
\code{\link{balance_ions}}
}
