% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biofilteration_model.R
\name{biofilter_toc}
\alias{biofilter_toc}
\title{Determine TOC removal from biofiltration}
\source{
Terry and Summers 2018
}
\usage{
biofilter_toc(water, ebct, o3_dose = 0)
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}. Water must include ph, doc, and uv254}

\item{ebct}{The empty bed contact time (min) used for the biofilter}

\item{o3_dose}{Applied ozone (O3) dose in mg/L. Defaults to 0 mg/L.}
}
\description{
This function applies the Terry model to a water created by \code{\link{define_water}} to determine biofiltered
DOC (mg/L).
}
\examples{
example1 <- define_water(ph = 7, temp = 15, alk = 100, toc = 4, doc = 3.8, uv254 = .1) \%>\%
  biofilter_toc(ebct = 20)
example1@toc

example2 <- define_water(ph = 7, temp = 15, alk = 100, toc = 4, doc = 3.8, uv254 = .1) \%>\%
  biofilter_toc(ebct = 20, o3_dose = 3)

}
