% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_base_equilibrium.R
\name{blend_waters}
\alias{blend_waters}
\title{Blend water function}
\usage{
blend_waters(
  water1,
  ratio1,
  water2,
  ratio2,
  water3 = data.frame(ph = NA),
  ratio3 = 0,
  water4 = data.frame(ph = NA),
  ratio4 = 0,
  ...
)
}
\arguments{
\item{water1}{Source water 1 data frame created by \code{\link{define_water}}}

\item{ratio1}{Blend ratio of water 1. (Blend ratios must sum to 1)}

\item{water2}{Source water 2 data frame created by \code{\link{define_water}}}

\item{ratio2}{Blend ratio of water 2. (Blend ratios must sum to 1)}

\item{water3}{Source water 3 data frame created by \code{\link{define_water}}}

\item{ratio3}{Blend ratio of water 3. (Blend ratios must sum to 1)}

\item{water4}{Source water 4 data frame created by \code{\link{define_water}}}

\item{ratio4}{Blend ratio of water 4. (Blend ratios must sum to 1)}
}
\description{
This function takes up to 4 water data frames defined by \code{\link{define_water}} and outputs a new water data frame with updated ion balance and pH.
}
\examples{
blend_waters(water1, 0.5, water2, .05)

}
\seealso{
\code{\link{define_water}}
}
