% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{calculate_dic}
\alias{calculate_dic}
\title{Calculate dissolved inorganic carbon (DIC) from total carbonate}
\usage{
calculate_dic(water)
}
\arguments{
\item{water}{a water class object containing columns with all the parameters listed in \code{\link{define_water}}}
}
\value{
A numeric value for the calculated DIC.
}
\description{
This function takes a water class object defined by \code{\link{define_water}}
and outputs a DIC (mg/L).
}
\examples{

example_dic <- define_water(8, 15, 200) \%>\%
  calculate_dic()

}
\seealso{
\code{\link{define_water}}
}
