% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{calculate_dic_chain}
\alias{calculate_dic_chain}
\title{calculate_dic_chain}
\usage{
calculate_dic_chain(df, input_water = "Finished")
}
\arguments{
\item{water}{a water class object containing columns with all the parameters listed in \code{\link{define_water}}}
}
\value{
A numeric value for the calculated DIC.
}
\description{
Calculate dissolved inorganic carbon (DIC) from total carbonate

This function takes a water class object defined by \code{\link{define_water}}
and outputs a DIC (mg/L). This function should be used when calculating DIC for multiple waters.
}
\examples{

example_dic <- water_df \%>\%
define_water_chain() \%>\%
calculate_dic_chain()


}
\seealso{
\code{\link{define_water}}
}
