% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorine_chloramine_decay.R
\name{chemdose_cl2}
\alias{chemdose_cl2}
\title{Calculate total OCL}
\source{
Ct, raw: U.S. EPA (2001) equation 5-113

Ct, coagulated/treated: U.S. EPA (2001) equation 5-117

Ct corrected: U.S. EPA (2001) equation 5-118

CAt: U.S. EPA (2001) equation 5-120

See references list at: \url{https://github.com/BrownandCaldwell/tidywater/wiki/References}
}
\usage{
chemdose_cl2(
  water,
  cl2_dose,
  time,
  treatment = "raw",
  cl_type = "chlorine",
  chlorine_correction = FALSE
)
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}}

\item{cl2_dose}{Applied chlorine or chloramine dose (mg/L as cl2_dose). Model results are valid for doses between 0.995 and 41.7 mg/L for raw water,
and for doses between 1.11 and 24.7 mg/L for coagulated water.}

\item{time}{Reaction time (hours). Chlorine decay model results are valid for reaction times between 0.25 and 120 hours.Chloramine decay model
does not have a time restriction.}

\item{treatment}{Type of treatment applied to the water. Options include "raw" for no treatment (default), "coag" for
water that has been coagulated or softened.}

\item{cl_type}{Type of chlorination applied, either "chlorine" (default) or "chloramine".}

\item{chlorine_correction}{Defaults to FALSE. Corrects the initial chlorine residual concentration when the cl_type is also "chlorine".}
}
\value{
An updated disinfectant residual in the tot_ocl water slot.
}
\description{
\code{chemdose_cl2} calculates the decay of chlorine/chloramine based on the U.S. EPA's
Water Treatment Plant Model (U.S. EPA, 2001). Required arguments include an object of class "water"
created by \code{\link{define_water}} chlorine/chloramine dose, type, reaction time, and treatment applied (if any).
The function also requires additional water quality parameters defined in \code{define_water}
including TOC and UVA.
}
\details{
The function will calculate haloacetic acids (HAA) as HAA5, and total trihalomethanes (TTHM).
The function returns a new object of class "water" with predicted chlorine/chloramine residual concentrations, which
is total/combined chlorine concentration. For free concentration, the chemdose_ph functions should be implemented after
this function.
Use \code{summarise_wq} to quickly tabulate the results.
}
\examples{
example_cl2 <- suppressWarnings(define_water(8, 20, 66, toc = 4, uv254 = .2, br = 50)) \%>\%
  chemdose_cl2(cl2_dose = 2, time = 8)
example_cl2 <- suppressWarnings(define_water(7.5, 20, 66, toc = 4, uv254 = .2, br = 50)) \%>\%
  chemdose_cl2(cl2_dose = 3, time = 168, treatment = "coag", cl_type = "chloramine")

}
