% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbp_models.R
\name{chemdose_dbp}
\alias{chemdose_dbp}
\title{Calculate DBP formation using dosed chlorine, ambient bromide, and reaction time.}
\usage{
chemdose_dbp(water, cl2, time, water_type = "treated")
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}}

\item{cl2}{Applied chlorine dose (mg/L as Cl2). Dose should be between 1.51 and 33.55 mg/L}

\item{time}{Reaction time (hours). Reaction time should be between 2 and 168 hours}

\item{water_type}{Type of treatment applied to the water. Default type is "treated", but
user may also specify "untreated". Untreated water is generally raw water. "Treated" water applies to
water that has been coagulated or softened.}
}
\description{
\code{chemdose_dbp} calculates disinfection byproduct (DBP) formation based on the WTP model. Parameters required
include chlorine addition, bromide, TOC, UV254, temperature, pH, and reaction time. This function will
calculate haloacetic acids (HAA) as HAA5, HAA6, or HAA9, as well as total trihalomethanes (TTHM).
The function takes an object of class "water" created by \code{\link{define_water}} and user-specified
chlorine addition, bromide, and reaction time and returns a dataframe of predicted DBP formation.
}
\details{
TTHMs, untreated: Amy et al. (1998), WTP Model v. 2.0, equation 5-131
HAAs, untreated: Amy et al. (1998), WTP Model v. 2.0, equation 5-134

TTHMs, treated: Amy et al. (1998), WTP Model v. 2.0, equation 5-139
HAAs, treated: Amy et al. (1998), WTP Model v. 2.0, equation 5-142
}
\examples{
example_dbp <- suppressWarnings(define_water(7.5, 20, 66, toc = 4, uv254 = .2, br = 50)) \%>\%
chemdose_dbp(cl2 = 2, time = 8)
example_dbp <- suppressWarnings(define_water(8, 25, 66, toc = 4, uv254 = .2, br = 50)) \%>\%
chemdose_dbp(cl2 = 2, time = 8, water_type = "untreated")

}
