% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_base_equilibirum.R
\name{chemdose_ph}
\alias{chemdose_ph}
\title{Chemical Dose Function}
\usage{
chemdose_ph(
  water,
  hcl = 0,
  h2so4 = 0,
  h3po4 = 0,
  naoh = 0,
  na2co3 = 0,
  nahco3 = 0,
  caoh2 = 0,
  mgoh2 = 0,
  cl2 = 0,
  naocl = 0,
  caocl2 = 0,
  co2 = 0,
  alum = 0,
  fecl3 = 0,
  fe2so43 = 0
)
}
\arguments{
\item{water}{Source water of class "water" created by \code{\link{define_water}}}

\item{hcl}{Hydrochloric acid: HCl -> H + Cl}

\item{h2so4}{Sulfuric acid: H2SO4 -> 2H + SO4}

\item{h3po4}{Phosphoric acid: H3PO4 -> 3H + PO4}

\item{naoh}{Caustic: NaOH -> Na + OH}

\item{na2co3}{Soda ash: Na2CO3 -> 2Na + CO3}

\item{nahco3}{Sodium bicarbonate: NaHCO3 -> Na + H + CO3}

\item{caoh2}{Lime: Ca(OH)2 -> Ca + 2OH}

\item{mgoh2}{Magneisum hydroxide: Mg(OH)2 -> Mg + 2OH}

\item{cl2}{Chlorine gas: Cl2(g) + H2O -> HOCl + H + Cl}

\item{naocl}{Sodium hypochlorite: NaOCl -> Na + OCl}

\item{caocl2}{Calcium hypochlorite: Ca(OCl)2 -> Ca + 2OCl}

\item{co2}{Carbon Dioxide CO2 (gas) + H2O -> H2CO3*}

\item{alum}{Hydrated aluminum sulfate Al2(SO4)3*14H2O + 6HCO3 -> 2Al(OH)3(am) +3SO4 + 14H2O + 6CO2}

\item{fecl3}{Ferric Chloride FeCl3 + 3HCO3 -> Fe(OH)3(am) + 3Cl + 3CO2}

\item{fe2so43}{Ferric sulfate Fe2(SO4)3 + 6HCO3 -> 2Fe(OH)3(am) +3SO4 + 6CO2}
}
\description{
This function takes chemical doses and a water data frame defined by \code{\link{define_water}} and outputs a new water data frame with updated ion balance and pH.
Units of all chemical additions in mg/L as chemical (not as product).
Returns "water" class object of dosed water.
}
\examples{
# Put example code here

}
\seealso{
\code{\link{define_water}}
}
