% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc_removal.R
\name{chemdose_toc}
\alias{chemdose_toc}
\title{Determine TOC removal from coagulation}
\usage{
chemdose_toc(water, alum = 0, fecl3 = 0, fe2so43 = 0, coeff = "Alum")
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}. Water must include ph, doc, and uv254}

\item{alum}{Amount of hydrated aluminum sulfate added in mg/L: Al2(SO4)3*14H2O + 6HCO3 -> 2Al(OH)3(am) +3SO4 + 14H2O + 6CO2}

\item{fecl3}{Amount of ferric Chloride added in mg/L: FeCl3 + 3HCO3 -> Fe(OH)3(am) + 3Cl + 3CO2}

\item{fe2so43}{Amount of ferric sulfate added in mg/L: Fe2(SO4)3 + 6HCO3 -> 2Fe(OH)3(am) +3SO4 + 6CO2}

\item{coeff}{String specifying the Edwards coefficients to be used from "Alum", "Ferric", "General Alum", "General Ferric", or "Low DOC" or
named vector of coefficients, which must include: k1, k2, x1, x2, x3, b}
}
\description{
This function applies the Edwards (1997) model to a water created by \code{\link{define_water}} to determine coagulated 
DOC. Coagulated UVA is from the Water Treatment Plant Model. Note that the models rely on pH of coagulation. If pH is 
unknown, utilize \code{\link{chemdose_ph}} first.
}
\examples{
# example code


}
