% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_water}
\alias{convert_water}
\title{Convert Water}
\usage{
convert_water(water)
}
\arguments{
\item{water}{A "water" class object}
}
\description{
Function to convert from Water class to a dataframe.
This is useful for one-off checks. It may be better to use fn_once
}
\examples{

library(tidyverse)

# Generates 1 row dataframe
example_df <- define_water(ph = 7, temp = 20, alk = 100) \%>\%
convert_water()

example_df <- water_df \%>\%
define_water_chain() \%>\%
mutate(to_dataframe  = map(defined_water, convert_water)) \%>\%
unnest(to_dataframe) \%>\%
select(-defined_water)

}
\seealso{
\code{\link{define_water}}
}
