% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{define_water}
\alias{define_water}
\title{Create water class object given water quality parameters}
\usage{
define_water(
  ph,
  temp,
  alk,
  tot_hard,
  ca_hard,
  na,
  k,
  cl,
  so4,
  tot_ocl = 0,
  tot_po4 = 0
)
}
\arguments{
\item{ph}{water pH}

\item{temp}{Temperature in degree C}

\item{alk}{Alkalinity in mg/L as CaCO3}

\item{tot_hard}{Total hardness in mg/L as CaCO3}

\item{ca_hard}{Calcium hardness in mg/L as CaCO3}

\item{na}{Sodium in mg/L Na+}

\item{k}{Potassium in mg/L K+}

\item{cl}{Chloride in mg/L Cl-}

\item{so4}{Sulfate in mg/L SO42-}

\item{tot_ocl}{Chlorine in mg/L as Cl2. Used when a starting water has a chlorine residual.}

\item{tot_po4}{Phosphate in mg/L as PO4. Used when a starting water has a phosphate residual.}
}
\description{
This function takes water quality parameters and creates an S4 "water" class object that forms the input and output of all tidywater models.
Carbonate balance is calculated and units are converted to mol/L. Missing values are handled by defaulting to 0 or
NA. Calcium hardness defaults to 65% of the total hardness because that falls within a typical range. For best results
manually specify all water quality parameters in the define water arguments.
}
\examples{
water_missingions <- define_water(ph = 7, temp = 15, alk = 100)
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)

}
