% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_general.R
\name{define_water}
\alias{define_water}
\title{Define water vector}
\usage{
define_water(
  ph,
  temp,
  alk,
  tot_hard,
  ca_hard,
  na,
  k,
  cl,
  so4,
  tot_ocl = 0,
  po4 = 0
)
}
\arguments{
\item{ph}{water pH}

\item{temp}{Temperature in degree C}

\item{alk}{Alkalinity in mg/L as CaCO3}

\item{tot_hard}{Total hardness in mg/L as CaCO3}

\item{ca_hard}{Calcium hardness in mg/L as CaCO3}

\item{na}{Sodium in mg/L Na+}

\item{k}{Potassium in mg/L K+}

\item{cl}{Chloride in mg/L Cl-}

\item{so4}{Sulfate in mg/L SO42-}

\item{tot_ocl}{Chlorine in mg/L as ??}

\item{po4}{Phosphate in mg/L as PO4}
}
\description{
This function takes water quality parameters and creates a "water" object that forms the input and output of all pH functions.
Carbonate balance is calculated and units are converted to mol/L
}
\examples{
# Put example code here

}
