% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{define_water_chain}
\alias{define_water_chain}
\title{Define Water chain}
\usage{
define_water_chain(df, output_water = "defined_water")
}
\arguments{
\item{df}{a data frame containing columns with all the parameters listed in \code{\link{define_water}}}

\item{output_water}{name of the output column storing updated parameters with the class, Water. Default is "defined_water".}
}
\description{
This function allows \code{\link{define_water}} to be added to a piped data frame. 
Its output is a Water class, and can therefore be chained with "downstream" tidywater functions.
}
\examples{

example_df <- water_df \%>\% 
define_water_chain() \%>\% 
balance_ions_once()

example_df <- water_df \%>\% 
define_water_chain(output_water = "This is a column of water") \%>\% 
balance_ions_once()

}
\seealso{
\code{\link{define_water}}
}
