% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_cu.R
\name{dissolve_cu}
\alias{dissolve_cu}
\title{Calculate Dissolved Copper Concentration}
\source{
Lytle et al (2018)
}
\usage{
dissolve_cu(water, dic)
}
\arguments{
\item{water}{Source water object of class "water" created by \code{\link{define_water}}. Water must include ph and dic}
}
\value{
\code{dissolve_cu} returns a column containing dissolved copper concentration in mg/L.
}
\description{
This function takes a water defined by defined_water and output a column of dissolved copper.
For a single water, use \code{dissolve_cu}; to apply the model to a dataframe use \code{dissolve_cu_chain}.
}
\details{
Dissolved copper is a function of pH, DIC, and PO4. Output units are in mg/L. You may need to calculate DIC upsteam of this function
using calculate_dic.
For large datasets, using \code{fn_once} or \code{fn_chain} may take many minutes to run. These types of functions use the furrr package
for the option to use parallel processing and speed things up. To initialize parallel processing, use
\code{plan(multisession)} or \code{plan(multicore)} (depending on your operating system) prior to your piped code with the
\code{fn_once} or \code{fn_chain} functions. Note, parallel processing is best used when your code block takes more than a minute to run,
shorter run times will not benefit from parallel processing.
}
\examples{

example_cu <- define_water(ph = 7.5, alk = 125) \%>\%
  dissolve_cu()
}
