% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid Base Equilibrium.R
\name{dose}
\alias{dose}
\title{Chemical Dose Function}
\usage{
dose(
  water,
  hcl = 0,
  h2so4 = 0,
  h3po4 = 0,
  naoh = 0,
  na2co3 = 0,
  nahco3 = 0,
  caoh2 = 0,
  mgoh2 = 0,
  cl2 = 0,
  naocl = 0,
  caocl2 = 0
)
}
\arguments{
\item{water}{Source water data frame created by waterdef}

\item{hcl}{Hydrochloric acid: HCl -> H + Cl}

\item{h2so4}{Sulfuric acid: H2SO4 -> 2H + SO4}

\item{naoh}{Caustic: NaOH -> Na + OH}

\item{na2co3}{Soda ash: Na2CO3 -> 2Na + CO3}

\item{nahco3}{Sodium bicarbonate: NaHCO3 -> Na + H + CO3}

\item{caoh2}{Lime: Ca(OH)2 -> Ca + 2OH}

\item{mgoh2}{Magneisum hydroxide: Mg(OH)2 -> Mg + 2OH}

\item{cl2}{Chlorine gas: Cl2(g) + H2O -> HOCl + H + Cl}

\item{naocl}{Sodium hypochlorite: NaOCl -> Na + OCl}

\item{caocl2}{Calcium hypochlorite: Ca(OCl)2 -> Ca + 2OCl}

\item{h3op4}{Phosphoric acid: H3PO4 -> 3H + PO4}
}
\description{
This function takes chemical doses and a water data frame defined by 'waterdef' and outputs a new water data frame with updated ion balance and pH.
Units of all chemical additions in mg/L.
Returns data frame of dosed water quality.
}
\examples{
# Put example code here

}
\seealso{
waterdef
}
