% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{leadsol_constants}
\alias{leadsol_constants}
\title{Data frame of equilibrium constants for lead and copper solubility}
\format{
A dataframe with 1 row and 38 columns

Solids:
\describe{
\item{K_solid_lead_hydroxide}{Solid lead hydroxide}
\item{K_solid_cerussite}{Solid cerussite}
\item{K_solid_hydrocerussite}{Solid hydrocerussite}
\item{K_solid_hydroxypyromorphite_schock}{Solid hydroxypyromorphite, from Schock et al. 1996}
\item{K_solid_hydroxypyromorphite_zhu}{Solid hydroxypyromorphite, from Zhu et al. 2015}
\item{K_solid_pyromorphite_xie}{Solid pyromorphite, from Xie & Giammar 2007}
\item{K_solid_pyromorphite_topolska}{Solid pyromorphite, from Topolska et al. 2016}
\item{K_solid_primary_lead_ortho}{Solid primary lead orthophosphate}
\item{K_solid_secondary_lead_ortho}{Solid secondary lead orthophosphate}
\item{K_solid_tertiary_lead_ortho}{Solid tertiary lead orthophosphate}
\item{K_solid_anglesite}{Solid anglesite}
\item{K_solid_laurionite_nl}{Solid laurionite, from Nasanen & Lindell 1976}
\item{K_solid_laurionite_loth}{Solid laurionite, from Lothenbach et al. 1999}
}



Lead-Chloride Complexes:
\describe{
\item{K_1_Cl}{PbCl+1}
\item{B_2_Cl}{PbCl2}
\item{B_3_Cl}{PbCl3-}
\item{B_4_Cl}{PbCl4-2}
}

Sulfate Acid-Base Chemistry and Lead-Sulfate Complexes:
\describe{
\item{K_s}{SO4-2}
\item{K_1_SO4}{PbSO4}
\item{B_2_SO4}{Pb(SO4)2-2}
}

Carbonate Acid-Base Chemistry and Lead-Carbonate Complexes:
\describe{
\item{K_c_1}{Bicarbonate}
\item{K_c_2}{Carbonate}
\item{K_1_CO3}{PbHCO3+}
\item{K_2_CO3}{PbCO3}
\item{K_3_CO3}{Pb(CO3)2-2}
}

Phosphate Acid-Base Chemistry and Lead-Phosphate Complexes:
\describe{
\item{K_p_1}{H2PO4—}
\item{K_p_2}{HPO4—2}
\item{K_p_3}{PO4-3}
\item{K_1_PO4}{PbHPO4}
\item{K_2_PO4}{PbH2PO4+}
}
}
\source{
Benjamin, M. M. (2002) Water Chemistry, 1st Edition, McGraw-Hill, New York, NY.

Lothenbach, B., Ochs, M., Wanner, H. & Yui, M. (1999) Thermodynamic Data for the Speciation and Solubility of Pd, Pb, Sn, Sb, Nb and Bi in Aqueous Solution. Japan Nuclear Cycle Development Institute, Ibaraki, Japan.

Nasanen, R. & Lindell, E. (1976) Studies on Lead(II) Hydroxide Salts. Part I. The Solubility Product of Pb(OH)Cl, Finnish Chemical Letters, 95.

Powell, K.J., Brown, P.L., Byrne, R.H., Gajda, T., Hefter, G., Leuz, A.K., Sjoberg, S. & Wanner, H. (2009) Chemical Speciation of Environmentally Significant Metals with Inorganic Ligands - Part 3: The Pb2+, OH—, Cl—, CO32—, SO42—, and PO43— Systems - (IUPAC Technical Report). Pure and Applied Chemistry, 81:12:2425.

Powell, K.J., Brown, P.L., Byrne, R.H., Gajda, T., Hefter, G., Sjoberg, S. & Wanner, H. (2005) Chemical Speciation of Environmentally Significant Heavy Metals with Inorganic Ligands - Part 1: The Hg2+, Cl—, OH—, CO32—, SO42—, and PO43— Aqueous Systems - (IUPAC Technical Report). Pure and Applied Chemistry, 77:4:739.

Schock, M.R., Wagner, I. & Oliphant, R.J. (1996) Chapter 4 - Corrosion and Solubility of Lead in Drinking Water. Internal Corrosion of Water Distribution Systems, 2nd Edition. American Water Works Association Research Foundation, Denver, CO.

Topolska, J., Manecki, M., Bajda, T., Borkiewicz, O. & Budzewski, P. (2016) Solubility of Pyromorphite Pb5(PO4)3Cl at 5-65 °C and Its Experimentally Determined Thermodynamic Parameters. The Journal of Chemical Thermodynamics, 98:282.

Xie, L. & Giammar, D.E. (2007) Equilibrium Solubility and Dissolution Rate of the Lead Phosphate Chloropyromorphite. Environmental Science & Technology, 41:23:8050.

Zhu, Y.N., Zhu, Z.Q., Zhao, X., Liang, Y.P. & Huang, Y.H. (2015) Characterization, Dissolution, and Solubility of Lead Hydroxypyromorphite [Pb5(PO4)3OH] at 25-45 °C. Journal of Chemistry, 2015:269387:1.

Wahman, D. G., Pinelli, M. D., Schock, M. R., & Lytle, D. A. (2021). Theoretical equilibrium lead(II) solubility revisited: Open source code and practical relationships. AWWA Water Science, e1250. https://doi.org/10.1002/aws2.1250
}
\usage{
leadsol_constants
}
\description{
A dataset containing equilibrium constants for lead and copper solubility
}
\keyword{datasets}
