% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{leadsol_constants}
\alias{leadsol_constants}
\title{Data frame of equilibrium constants for lead and copper solubility}
\format{
A dataframe with 38 rows and 3 columns

Solids:
\describe{
\item{species_name}{Name of lead solid or complex with possible _letter to cite different references}
\item{constant_name}{Reference ID for constants}
\item{log_value}{Equilibrium constant log value}
\item{source}{Source for equilibrium constant value}
}
}
\source{
Benjamin (2010)

Lothenbach et al. (1999)

Nasanen & Lindell (1976)

Powell et al. (2009)

Powell et al. (2005)

Schock et al. (1996)

Topolska et al. (2016)

Xie & Giammar (2007)

Zhu et al. (2015)

Wahman et al. (2021)

See references list at: \url{https://github.com/BrownandCaldwell-Public/tidywater/wiki/References}
}
\usage{
leadsol_constants
}
\description{
A dataset containing equilibrium constants for lead solubility
}
\keyword{datasets}
