% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{mweights}
\alias{mweights}
\title{Molar weights of relevant compounds}
\format{
A dataframe with one row and one column per compound
}
\usage{
mweights
}
\description{
A dataset containing the molar weights of several compounds in g/mol. Column names are lowercase chemical formulas (with no charge),
with the exception of the following coagulants:
alum = Al2(SO4)3\emph{14H2O,
ferricchloride = FeCl3,
ferricsulfate = Fe2(SO4)3}8.8H2O,
}
\keyword{datasets}
