% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plot_ions}
\alias{plot_ions}
\title{Create summary plot of ions from water class}
\usage{
plot_ions(water, title = "")
}
\arguments{
\item{water}{Source water vector created by link function here}

\item{title}{Optional plot title}
}
\description{
This function takes a water data frame defined by \code{\link{define_water}} and outputs an ion balance plot.
}
\examples{
water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)
plot_ions(water_defined)

}
