% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plot_lead}
\alias{plot_lead}
\title{Create dissolved lead and DIC contour plot given input data frame}
\usage{
plot_lead(df)
}
\arguments{
\item{df}{Source data as a data frame. Must have columns for pH and dic data. Columns containing
a single temperature and TDS must also be included.}
}
\value{
A ggplot object displaying a contour plot of dissolved lead, pH, and DIC
}
\description{
This function takes a data frame and outputs a contour plot of dissolved lead and DIC plot.
}
\examples{
historical <- data.frame(ph = c(7.7, 7.86, 8.31, 7.58, 7.9, 8.06, 7.95, 8.02, 7.93, 7.61),
                         dic = c(14.86, 16.41, 16.48, 16.63, 16.86, 16.94, 17.05, 17.23, 17.33, 17.34), 
                         temp = 25,
                         tds = 200)
plot_lead(historical)

}
