% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_base_equilibirum.R
\name{solvedose_ph}
\alias{solvedose_ph}
\title{Target Chemical Dose Function}
\usage{
solvedose_ph(water, target_ph, chemical)
}
\arguments{
\item{water}{Source water of class "water" created by \code{\link{define_water}}}

\item{target_ph}{The final pH to be achieved after the specified chemical is added.}

\item{chemical}{The chemical to be added. Current supported chemicals include:
acids: "hcl", "h2so4", "h3po4", "co2"; bases: "naoh", "na2co3", "nahco3", "caoh2", "mgoh2"}
}
\description{
This function calculates the required amount of a chemical to dose based on a target pH and existing water quality.
Returns numeric value for dose in mg/L. Uses uniroot on the \code{\link{chemdose_ph}} function.
}
\examples{
# Put example code here

}
\seealso{
\code{\link{define_water}}
}
