% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_base_equilibirum.R
\name{summarize_wq}
\alias{summarize_wq}
\title{Water Summary Table}
\usage{
summarize_wq(water)
}
\arguments{
\item{water}{Source water vector created by link function here}
}
\description{
This function takes a source water vector and outputs a formatted summary table.
}
\examples{
# Put example code here

}
