% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{package}
\name{tidywater-package}
\alias{tidywater}
\alias{tidywater-package}
\title{tidywater: Water Quality Models for Drinking Water Treatment Processes}
\description{
Provides multiple water chemistry-based models and published empirical models in one standard format. Functions can be chained together to model a complete treatment process and are designed to work in a 'tidyverse' workflow. Models are primarily based on these sources: Benjamin, M. M. (2002) Water Chemistry, 1st Edition, McGraw-Hill, New York, NY. Crittenden, J. C., Trussell, R., Hand, D., Howe, J. K., & Tchobanoglous, G., Borchardt, J. H. (2012). MWH’s Water Treatment: Principles and Design (Third). John Wiley & Sons Inc. USEPA. (2001). Water Treatment Plant Model, Version 2.0, User’s Manual. \url{https://www.epa.gov/sites/default/files/2017-03/documents/wtp_model_v._2.0_manual_508.pdf}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BrownandCaldwell-Public/tidywater}
  \item Report bugs at \url{https://github.com/BrownandCaldwell-Public/tidywater/issues}
}

}
\author{
\strong{Maintainer}: Sierra Johnson \email{sjohnson2@brwncald.com}

Authors:
\itemize{
  \item Libby McKenna \email{lmckenna@brwncald.com}
  \item Riley Mulhern \email{rmulhern@brwncald.com} (\href{https://orcid.org/0000-0001-6293-3672}{ORCID})
  \item Chris Corwin \email{ccorwin@brwncald.com} (\href{https://orcid.org/0000-0002-9462-0352}{ORCID})
}

Other contributors:
\itemize{
  \item Brown and Caldwell [funder, copyright holder]
}

}
\keyword{internal}
