% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{package}
\name{tidywater-package}
\alias{tidywater}
\alias{tidywater-package}
\title{tidywater: Water quality modeling for water treatment}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
