% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywater-package.R
\docType{data}
\name{vlog_removalcts}
\alias{vlog_removalcts}
\title{Data frame of virus log removal CT values}
\format{
A dataframe with 35 rows and 4 columns
\describe{
\item{ph_range}{Specifies PAC type}
\item{A}{Constant in the PAC model}
\item{a}{Coefficient in PAC model, associated with DOC0}
\item{b}{Coefficient in PAC model, associated with dose}
\item{c}{Coefficient in PAC model, associated with time}
}
}
\source{
Cho (2007)
}
\usage{
vlog_removalcts
}
\description{
A dataset containing coefficients for calculating PAC TOC removal
}
\keyword{datasets}
